Added a question store with hard coded questions and added a call to it in the message backend button. Some code clean up as well
