+++ Assets/components/Managers.prefab
@@ -77,6 +77,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   playerScoring: {fileID: 1274120691537752878}
+  questionStore: {fileID: 3447547822474969686}
 --- !u!1 &3447547823533500156
 GameObject:
   m_ObjectHideFlags: 0
@@ -129,6 +130,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 3447547824102527054}
+  - component: {fileID: 3447547822474969686}
   m_Layer: 0
   m_Name: QuestionStore
   m_TagString: Untagged
@@ -150,3 +152,15 @@ Transform:
   m_Father: {fileID: 3447547822414470340}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &3447547822474969686
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3447547824102527053}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9682d9525b8b52346a2593376765000f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
