+++ Assets/src/Cat1200.cs
@@ -1,0 +1,24 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class Cat1200 : MonoBehaviour
+{
+    public Button cat1200;
+    // Start is called before the first frame update
+    void Start()
+    {
+        cat1200.onClick.AddListener(TaskOnClick);
+    }
+    void TaskOnClick()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
